requirements.txt
pillow==11.3.0
opencv-python==4.11.0.86
numpy==2.3.1
pdf2image==1.17.0
pytesseract==0.3.13
PySide6==6.9.1
python-docx==1.2.0
